+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.4
+- Added option to automatically open GUI on death
+  - Either it can show all chests or it can directly show the GUI for latest chest
+  - Option to only open the GUI when player is allowed to fetch or teleport to the chest added as well
+
 ## 0.0.3
 - Fixed /acunlock command and unlocking through GUI
 - Only show buttons in GUI that the player has permission to